Rename package commands to command
